--- client/src/components/__tests__/ShareSocialMedia.test.js
@@ -1,21 +1,0 @@
-/**
- * @jest-environment jsdom
- */
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import ShareSocialMedia from '../ProductDetail/ShareSocialMedia';
-
-describe('Social Media Icons', () => {
-  test('Should render the Facebook icon', () => {
-    render(<ShareSocialMedia />);
-    expect(screen.getByTestId('fb')).not.toEqual(null);
-  });
-  test('Should render the Twitter icon', () => {
-    render(<ShareSocialMedia />);
-    expect(screen.getByTestId('twt')).not.toEqual(null);
-  });
-  test('Should render the Pinterest icon', () => {
-    render(<ShareSocialMedia />);
-    expect(screen.getByTestId('pnt')).not.toEqual(null);
-  });
-});
